finish courier details screen

+++ app/src/main/java/com/nat/couriersapp/screens/courierDetails/domain/models/CourierBody.kt
@@ -2,5 +2,5 @@ package com.nat.couriersapp.screens.courierDetails.domain.models
 
 data class CourierBody(
     val WaybillId : Int,
-    val WaybillSerial : Int
+    val WaybillSerial : Long
 )
